calculating computer best move is another class responsibility
